requirements.txt
albumentations==2.0.8
hydra-core==1.3.2
loguru==0.7.2
matplotlib==3.10.3
numpy==2.1.1
protobuf==6.32.0
omegaconf==2.3.0
onnx==1.17.0
onnxscript==0.5.4
onnxconverter-common==1.15.0


onnxruntime==1.21.1

onnxsim==0.4.36
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84


openvino==2025.2.0

pandas==2.2.3
Pillow==11.1.0
pillow_heif==0.18.0
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.15.1


# tensorrt==10.10.0.31

torchmetrics==1.7.4
faster-coco-eval==1.6.5
pycocotools==2.0.8
tqdm==4.66.5
difPy==4.2.1
gradio==5.37
wandb==0.21.0
tabulate==0.9.0

torch==2.9.0
torchvision==0.24.0

